Issue #3413 Separate the ringFragmentSource management from EVBC
The same interface is provided by the EVBC but the preferred mechanism 
for setting up ringFragmentSources is via the ::EVBC::registerRingSource 
proc rather than the startEVBSources. It is still possible to use 
startEVBSources and both paradigms can work at the same time. The 
management is now under the ::RingSourceMgr bundle. All of the previous
logic that deal with launching sources is actually embodied in the 
new ::RingSourceMgr bundle and most calls to ::EVBC that deal with 
ringFragmentSources are just delegated to the new bundle.