Issue #4924 Adds a file analyzer to the offlineorderer
This commit adds a tool to the offlineorderer that really reduces
the up front knowledge required for setting up a job. The user
can analyze their file using a new program called FileAnalyzer that
is a very simple filter implementation and the results of the
analysis are used to fill in as many details about their run as possible.

- It also adds the Process snit::type to the tclutils directory which 
  is a simple mechanism for handling pipeline processes.

- It also fixes the problem of build windows in the event builder when
  cat'ing files to it. The 20 second default has been observed to be too
  small in the past and this sets it to 600 seconds in the current scheme.