Fix error in driver skeletons
The driver skeletons had a makefile includes that wound up having two
instances of -lieee there's a flaw in libieee.so that makes a doubly
defined strong symbol if you do this.  That makes the builds of user
drivers fail.   The origin of the problem from our end was using
TCL_LDFLAGS rather than TCL_LIBS.