Fix error in create:
The branch of code that executed to see if an existing shared memory
was known to the ringmaster was faulty.  I must analyze the return data
to see if the ringbuffer name is present e.g. for the ring foxy the string

"{foxy "

is present in the usage string if that ring exists and is known to the
ring master.