Issue #159 - When used directly in a pipline Glom fails to write last items.
Specifically observed to miss the end run item.. but could be more
depending on timing.  This commit:
*  sets action.sa_handler to SIG_IGN in Os::blockSignal for good measure
(the signal was in the block mask already).
*  invokes Os::blockSignal(SIGPIPE) in glomMain.cpp:main

Note if glom's call to blockSignal fails, we output an error message
to stderr (never to stdout which is where ring items go) and continue
noting that last items could be omitted.