Issue #221 - readout wizard set title/run don't work.
Fixed the program name parameter which missed the _readout suffix
for the definition of those programs.
In /transition if attempting a transtion to the state we're already in,
just return without trying.  This may also deal with cases when the transition is attempting from multiple sources simultaneously though it does 'allow' a generally disallowed transition.